requirements.txt
pycryptodome >= 3.15.0
rlp >= 3.0.0
eth-hash >= 0.3.3
eth-keys >= 0.4.0
eth-typing >= 3.2.0
eth-utils >= 2.0.0